+++ inst/doc/fill-pkg-description.R
@@ -10,11 +10,22 @@ library(attachment)
 ## ---- eval=FALSE--------------------------------------------------------------
 #  att_amend_desc()
 
+## -----------------------------------------------------------------------------
+# Copy package in a temporary directory
+tmpdir <- tempdir()
+file.copy(system.file("dummypackage",package = "attachment"), tmpdir, recursive = TRUE)
+dummypackage <- file.path(tmpdir, "dummypackage")
+# browseURL(dummypackage)
+att_amend_desc(path = dummypackage, inside_rmd = TRUE)
+
 ## ---- eval=FALSE--------------------------------------------------------------
 #  # bookdown Imports are in Rmds
 #  imports <- c("bookdown", attachment::att_from_rmds("."))
 #  attachment::att_to_desc_from_is(path.d = "DESCRIPTION",
 #                                  imports = imports, suggests = NULL)
+
+## ---- eval=FALSE--------------------------------------------------------------
+#  att_from_description() %>% find_remotes()
 
 ## ---- eval=FALSE--------------------------------------------------------------
 #  att_from_namespace()
@@ -45,5 +56,5 @@ library(attachment)
 dummypackage <- system.file("dummypackage", package = "attachment")
 
 att_from_rscripts(path = file.path(dummypackage, "R"))
-att_from_rmds(path = file.path(dummypackage, "vignettes"))
+att_from_rmds(path = file.path(dummypackage, "vignettes"), inside_rmd = TRUE)